Rename to legend header
